01_jira-get-tickets/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
idna==3.4
jira==3.5.2
jira2markdown==0.3.6
Markdown==3.4.4
numpy==1.26.2
oauthlib==3.2.2
packaging==23.1
pandas==2.1.4
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.0.4
